[jit] kill script namespace
Once upon a time we thought this was necessary. In reality it is not, so
removing it.

For backcompat, our public interface (defined in `api/`) still has
typedefs to the old `script::` names.

There was only one collision: `Pass` as a `Stmt` and `Pass` as a graph
transform. I renamed one of them.

ghstack-source-id: 44fdadb045fe32803eaec6a29da056df6486babb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34515